requirements.txt
Django>=4.0.4
psycopg2>=2.8
celery>=5.2.6
redis>=4.3.1
flower>=1.0.0
django-celery-results>=2.3.1
djangorestframework>=3.13.1
PyJWT>=2.4.0
